Problem fixed on the teams page
