+++ src/pages/AdminPage/DBRightSection/Team/Profile/ProfileInfo.jsx
@@ -10,20 +10,20 @@ function ProfileInfo() {
           alt=""
           className="md:w-[110px] w-[90px] md:h-[110px] h-[90px]"
         />
+
         <div>
-          <h6 className="text-[#000000] font-Bold text-3xl">Jimmy Jules</h6>
+          <div className="flex gap-10">
+            <h6 className="text-[#000000] font-Bold text-3xl">Jimmy Jules</h6>
+            <button className="lg:hidden inline-block bg-[#DF327B] rounded-full px-2  text-white font-Bold">
+              <a href="" className="">
+                <FaPencilAlt />
+              </a>
+            </button>
+          </div>
           <p className="text-[#B44DB8]">jimmyjulle98@gmail.com</p>
         </div>
 
-        <button className="hidden lg:inline-flex bg-[#DF327B] rounded-full px-5 py-2 text-white font-Bold ml-auto mr-20">
-          <a href="" className="flex items-center">
-            <FaPencilAlt className="mr-2" />
-            Manage
-          </a>
-        </button>
-      </div>
-      <div className="">
-        <button className="lg:hidden inline-block bg-[#DF327B] rounded-full px-2 py-1.5 text-white font-Bold">
+        <button className="hidden lg:inline-flex bg-[#DF327B] rounded-full px-5 py-2 text-white font-Bold ml-auto lg:mr-20">
           <a href="" className="flex items-center">
             <FaPencilAlt className="mr-2" />
             Manage
@@ -31,7 +31,9 @@ function ProfileInfo() {
         </button>
       </div>
 
-      <DateInfor />
+      <div className="">
+        <DateInfor />
+      </div>
     </div>
   );
 }
